+++ Assets/Externals/2D Casual UI.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: eea07922793fb45d98e734985805ce9c
+folderAsset: yes
+timeCreated: 1486033941
+licenseType: Store
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
