+++ Assets/Externals/2D Casual UI/Original Artwork.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 340aa2cbeb80a4c48846c87c039e46a1
+folderAsset: yes
+timeCreated: 1486034034
+licenseType: Store
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
