+++ Assets/Externals/2D Casual UI/Sprite.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: e768493ddca804421b25bd6727f9c844
+folderAsset: yes
+timeCreated: 1486033989
+licenseType: Store
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
