+++ Assets/Externals/2D Casual UI/readme.rtf
@@ -1,0 +1,30 @@
+{\rtf1\ansi\ansicpg1252\cocoartf1504\cocoasubrtf810
+{\fonttbl\f0\fswiss\fcharset0 Helvetica;}
+{\colortbl;\red255\green255\blue255;}
+{\*\expandedcolortbl;;}
+\paperw11900\paperh16840\margl1440\margr1440\vieww10800\viewh15520\viewkind0
+\pard\tx566\tx1133\tx1700\tx2267\tx2834\tx3401\tx3968\tx4535\tx5102\tx5669\tx6236\tx6803\pardirnatural\partightenfactor0
+
+\f0\fs24 \cf0 HI THERE!\
+\
+Thank you so much for choosing our work. Just want to let you know that we are appreciated.\
+\
+This package contains graphic sources for 2D casual User Interface. HD image quality. One and done. Sliced sprite-sheet. Import package and you can use them right away. Great for using in any casual game or application.\
+\
+This package gives you\'85\
+\
+== Initial Artworks ==\
+- 28 icons\
+- Currencies\
+- Various styles of button\
+- Popup backgrounds\
+- Other elements\
+\
+PLUS, we give you the original PSD file !!!\
+Separated layer, easy to change colour. Allow you to adapt, re-design or create more.\
+\
+Please don\'92t forget to rate and comment on our work. Let us know if something is bad or need to be fixed. We also listening for your request if anything you want us to do next :\}\
+\
+Cheers,\
+mimustudio@gmail.com (Ploi)\
+www.mimu-studio.com}
