+++ Assets/Externals/2D Casual UI/readme.rtf.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: d3edae2e2c75b49ab9c5ee85393f2486
+timeCreated: 1486041554
+licenseType: Store
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
