+++ Assets/Scripts/SlotMachine.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6a992288974b55747803fc005fa8b0f6
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
